Refactor: improve stopwatch time handling and formatting
