+++ index.html
@@ -2,9 +2,9 @@
 <html>
 	<head>
 		<meta charset="UTF-8" />
-		<link rel="icon" type="image/svg+xml" href="/react.svg" />
+		<link rel="icon" type="image/svg+xml" href="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS7-NeG3mTTuuxgGhYzFz3wvyPNsh4ZfwGbqQ&s" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Rspack + React + TS</title>
+		<title>Stopwatch</title>
 	</head>
 	<body>
 		<div id="root"></div>
