+++ src/shared/consts/consts.ts
@@ -1,0 +1,3 @@
+export const SEC_IN_HOUR = 3600;
+export const MIN_IN_HOUR = 60;
+export const SEC_IN_MIN = 60;
